+++ CHANGELOG.md
@@ -15,6 +15,10 @@
 
 * Add utilities for emoji detection in messages and make all-emoji message larger font size ([#519](https://github.com/TryQuiet/quiet/issues/519))
 
+# Chores
+
+* Cleanup data directory at end of e2e tests
+
 [2.1.2]
 
 # Refactorings: